used bcrypt for hashing password for signup
in the user model a static method was added to handle the signup. Not only would this throw an error if an already existing email is used, but it will also properly hash the password with salt and then save it to the database. this static method will be used in the userController's signup method, within a try catch so that any errors can be caught. BTW, mongoose will already give an error when an existing email is going to be added because the email field was defined as unique in the schema, but in using this signup method we can control what type of error message will be transmitted as a response when the user enters an already existing password