added in Login Page and Signup Page
also fixed two issues, changed wrong keyword class to className in WorkoutDetils component for the styling of the delete button, also edited the template that would return a single workout detail in the map function found in the home page so that the child elements were just the <workoutdetails> tag and not wrapped in <></> because although the <workdetails> had a unique to each elment that was rendered when it was wrapped in <></> these entire element by itself didnot have an id. so leaving just the <workoutdetails> made sure that the child element that was rendered did indeed have an id. Also added some styles for the navBar, Login page, Signup Page and also implemented the login page and signup page in the App.js routes and also added the links in the nav bar