added new hook UseLogin and implemented it in Login page
ISSUE FIXED: there was an issue with type of request that /login was routed for. It was routed as a get request but the frontend server doesnot allow a body to be sent with a get request (although weirdly enough Postman allows a body to be sent with a GET request and the route worked just fine with Postman). so changed the login request to a post request so that the user login details can be sent to the backend as a part of the body.