added Api Route Protection
Created new middleware for the backend called RequireAuth.js. What this middle ware is used for is the protection of api routes by checking the users token validity and the then providing access to the normal workout routes. If a valid token is not present the workout routes cannot be accessed. Implmented this middleware in the workoutRoute.js